stable version with tests
