+++ public/js/app.js
@@ -9,7 +9,16 @@ window.addEventListener("DOMContentLoaded", () => {
 
 let maze = new Maze(10, [], 0, "");
 
+let charPos = `cell_1_1`;
+let charPosY = 1;
+let charPosX = 1;
+let steps = 0;
+
 function initMaze(dimensions) {
+  charPos = `cell_1_1`;
+  charPosY = 1;
+  charPosX = 1;
+  steps = 0;
   // Remove maze HTMLElements
   mazeFun.removeMaze();
   // Initialize maze state, create nodes and generate HTML view of maze
@@ -23,7 +32,48 @@ function initMaze(dimensions) {
   // Get image and place it at the end
   mazeFun.placeTarget();
   // Listen for arrow keys
-  document.onkeydown = mazeFun.moveCharacter;
+  document.onkeydown = moveCharacter;
+}
+
+function moveCharacter(e) {
+  e = e || window.event;
+  const character = document.getElementById("character");
+
+  if (e.keyCode == "38" && charPosY > 1 && $(`#${charPos}`).hasClass("no-top-border")) {
+    charPosY--;
+  } else if (e.keyCode == "40" && charPosY < maze.dimensions && $(`#${charPos}`).hasClass("no-bottom-border")) {
+    charPosY++;
+  } else if (e.keyCode == "37" && charPosX > 1 && $(`#${charPos}`).hasClass("no-left-border")) {
+    charPosX--;
+  } else if (e.keyCode == "39" && charPosX < maze.dimensions && $(`#${charPos}`).hasClass("no-right-border")) {
+    charPosX++;
+  }
+  steps++;
+  charPos = `cell_${charPosY}_${charPosX}`;
+
+  character.parentNode.removeChild(character);
+  document.getElementById(charPos).appendChild(character);
+  checkWinCondition(charPos);
+}
+
+function checkWinCondition(pos) {
+  if ($(`#${pos}`).hasClass("finish")) {
+    const character = document.getElementById("character");
+    character.parentNode.removeChild(character);
+    document.onkeydown = null;
+    const winPopup = document.getElementById("win-popup");
+    winPopup.style.display = "flex";
+    const p = document.createElement("p");
+    p.classList.add("row");
+    winPopup.appendChild(p);
+    pause();
+    p.innerHTML = `You took ${steps} steps and ${formattedTime} time to complete the map`;
+    reset();
+    maze.steps = steps;
+    maze.formattedTime = formattedTime;
+    userFun.updateDoc(maze);
+    userFun.newMazeToFirestore(maze);
+  }
 }
 
 function tooltip() {
@@ -139,4 +189,20 @@ $("#logout-btn").click(function () {
 });
 $("#match-history-btn").click(function () {
   userFun.showMatchHistory();
+});
+document.getElementById("closeButton").addEventListener(
+  "click",
+  function (e) {
+    e.preventDefault();
+    this.parentNode.style.display = "none";
+  },
+  false
+);
+document.addEventListener("click", () => {
+  const winPopup = document.getElementById("win-popup");
+  winPopup.style.display = "none";
+  const p = document.querySelector("#win-popup p");
+  if (p != null) {
+    winPopup.removeChild(p);
+  }
 });
